cast to unsigned

+++ test/common/http/utility_fuzz_test.cc
@@ -81,11 +81,13 @@ DEFINE_PROTO_FUZZER(const test::common::http::UtilityTestCase& input) {
   }
   case test::common::http::UtilityTestCase::kInitializeAndValidate: {
     const auto& options = input.initialize_and_validate();
+    // if there are duplicates, abort the test
     std::set<int> uniques;
     for (auto& setting : options.custom_settings_parameters()) {
       uniques.insert(setting.identifier().value());
     }
-    if (uniques.size() != options.custom_settings_parameters().size()) {
+    // protobuf repeated's size is a signed integer
+    if (uniques.size() != static_cast<uint32_t>(options.custom_settings_parameters().size())) {
       break;
     }
     Http2::Utility::initializeAndValidateOptions(options);
